Adds CI/CD pipeline configuration (#2)
* Adds CI/CD pipeline configuration

Sets up a CI/CD pipeline using GitHub Actions to automate the build, test, and package process for the application.

The pipeline is triggered on pushes to the `master` and `develop` branches, as well as on pull requests targeting these branches.

It includes jobs for building the Angular application, running tests, and packaging the application for deployment,
and uploads build artifacts to be used in subsequent deployment steps.

* Moves CI/CD workflow file

Updates the location of the CI/CD workflow file to adhere to standard GitHub Actions directory structure.

* Fix pipeline

* Pipeline Fix 2

* Pipeline Fix 3