Merge from develo to master (#4)
* Initializes Angular project with Tailwind CSS

Sets up a new Angular project with a modern toolchain.

- Integrates Tailwind CSS for styling.
- Configures EditorConfig for consistent code formatting.
- Includes basic VS Code configuration for enhanced development experience.
- Adds initial application structure and component.
- Includes expert recommendations to guide development.

* Moving from Karma to Jest as testing framework (#3)

* Removes Karma testing framework

Removes Karma and related dependencies from the project.
This change aims to simplify the testing setup.

* Integrates Jest testing framework

* Adds CI/CD pipeline configuration (#2)

* Adds CI/CD pipeline configuration

Sets up a CI/CD pipeline using GitHub Actions to automate the build, test, and package process for the application.

The pipeline is triggered on pushes to the `master` and `develop` branches, as well as on pull requests targeting these branches.

It includes jobs for building the Angular application, running tests, and packaging the application for deployment,
and uploads build artifacts to be used in subsequent deployment steps.

* Moves CI/CD workflow file

Updates the location of the CI/CD workflow file to adhere to standard GitHub Actions directory structure.

* Fix pipeline

* Pipeline Fix 2

* Pipeline Fix 3